tfrecord-builder/requirements.txt
Flask==1.0.2
google-cloud-pubsub==0.39.1
gunicorn==19.9.0
google-cloud-datastore
google-cloud-storage
google-api-python-client
numpy
tensorflow